Renamed models to model for package
