+++ src/main/java/uk/gov/companieshouse/bankruptofficersearch/api/model/response/ScottishBankruptOfficerSearchFiltersEntity.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.bankruptofficersearch.api.models.response;
+package uk.gov.companieshouse.bankruptofficersearch.api.model.response;
 
 import org.springframework.stereotype.Component;
