+++ src/main/java/uk/gov/companieshouse/bankruptofficersearch/api/model/rest/ScottishBankruptOfficerSearchResults.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.bankruptofficersearch.api.models.rest;
+package uk.gov.companieshouse.bankruptofficersearch.api.model.rest;
 
 import java.util.List;
